+++ src/components/organisms/StudentEditModal/StudentEditModal.tsx
@@ -1,6 +1,6 @@
 import { Checkbox, Label, ModalButton, SelectOption } from '@/components/atoms';
 import { InputForm } from '@/components/molecules';
-import { StudentType } from '@/services/admin/types';
+import { StudentType } from '@/types/admin/types';
 import { useEffect, useState } from 'react';
 import { useGenerationStore } from '@/stores/generationStore';
 import { convertSelectType } from '@/utils/convertSelectType';
