+++ src/components/templates/WebRecommendCards.tsx
@@ -28,7 +28,6 @@ const WebRecommendCards = ({
             </div>
         );
     }
-    console.log(data);
 
     return (
         <ul className="grid w-full grid-cols-4 gap-6">
