+++ src/services/common/generation/generationAPI.ts
@@ -1,5 +1,5 @@
+import { GenerationType } from '@/types/common/type';
 import { api } from '../axiosInstance';
-import { GenerationType } from './types';
 
 export const getGenerationListAPI = async (): Promise<GenerationType[]> => {
     const response = await api.get('/generation');
