+++ src/services/common/generation/index.ts
@@ -1,3 +1,2 @@
 export * from './generationService';
 export * from './generationAPI';
-export * from './types';
