--- src/services/common/generation/types.ts
@@ -1,10 +1,0 @@
-export interface GenerationType {
-    id: number;
-    name?: string;
-    active: boolean;
-}
-
-export interface SelectOptionType {
-    value: string | number;
-    label: string;
-}
