+++ src/services/study/web/index.ts
@@ -1,2 +1,1 @@
 export * from './webAPI';
-export * from './type';
