+++ src/types/common/type.ts
@@ -8,3 +8,13 @@ interface PageInfo {
     totalElements: number;
     totalPages: number;
 }
+export interface GenerationType {
+    id: number;
+    name?: string;
+    active: boolean;
+}
+
+export interface SelectOptionType {
+    value: string | number;
+    label: string;
+}
